Add files via upload
Goal: Analyze the sales data to identify trends, top-selling products, and revenue metrics for business decision-making.

Step 1: Data Exploration This is the stage where I had to familiarize myself with the dataset. This gives me the opportunity to understand what the data is all about, discover columns that require cleaning, and also to checkmate if the dataset has the required columns to help me answer necessary business case questions.

To begin with, I had to structure it by converting it into a table. Ctrl+T was used and I had to name the table Sales_dataset.

The table has 10 columns (Order ID, Product, Quantity Ordered, Price each, order date, purchase address, month, Sales City, and Hour) and 185950 rows.

Step 2: Data Preparation This is where I had to prepare the dataset to make it fit for analysis. All observations in the exploration stage will be taken care of here.

It was observed that the month column shows the month number. So, I had to rename it as a month number. The price of each and sales columns represent currency, so I had to introduce the $ sign.

The order date column has a datatype of date-time, noting that the dataset has a column called hour, I had to convert the order date to a short date from the home tab.

I used the text function to create the following additional columns from the order date so as to be able to answer the question on trends:

1. Year

2. Month

3. Week Day

4. To get the column called Sales Growth Rate, I used the formula = [(Current Sales — Previous Sales) / Previous Sales] x 100

Step 3: Data Analysis this is where I start answering the required questions. I used the pivot table and chart to draw up insights and create pictorial representations.

Below are the insights drawn from the project goal:

Having generated the insights via the Pivot table, I had to use the following charts:

1. Line chart to show both monthly and daily revenue trend

2. Column charts to show the top 10 bestselling products and revenue by state

3. Bar chart to represent monthly revenue growth rate.

Insights

It was accounted that December, October, and November had the highest profit followed by December generating the highest.

Finding out what days of the week generated the most sales, the analysis revealed that most of the sales were done on Tuesdays. Making Tuesday the best-selling day of the week.

MacBook Pro Laptop and iPhone are the most sold products among the top 10 best-selling products.

Products are mostly sold in San Francisco and Los Angeles thereby making these States the most revenue-generating States.

There was an increase in revenue growth rate in December. This was followed by October, April, and November with slight differences in growth.